+++ site-cookbooks/chort/metadata.rb
@@ -1,0 +1,8 @@
+name             'chort'
+maintainer       'Leonis0813'
+license          'All rights reserved'
+description      'Installs/Configures chort'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.0'
+depends          'nginx'
+depends          'rvm'